testing/mochitest/tests/test_bug360220.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="/tests/SimpleTest/test.css" type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=360220
-->
<window title="Mozilla Bug 360220"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <title>Test for Bug 360220</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>      

<body  xmlns="http://www.w3.org/1999/xhtml">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=360220">Mozilla Bug 360220</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
</pre>
</body>

<menulist id="menulist">
  <menupopup>
    <menuitem id="firstItem" label="foo" selected="true"/>
    <menuitem id="secondItem" label="bar"/>
  </menupopup>
</menulist>
<script class="testbody" type="application/javascript">
<![CDATA[

/** Test for Bug 360220 **/

var menulist = document.getElementById("menulist");
var secondItem = document.getElementById("secondItem");
menulist.selectedItem = secondItem;

ok(menulist.label == "bar", "second item was not selected");
secondItem.label = "new label";
ok(menulist.label == "new label", "menulist label was not updated to the label of its selected item");
]]>
</script>

</window>
